requirements.txt
streamlit
biopython
python-dotenv